requirements.txt
# UI Framework
streamlit>=1.20.0

# Data Analysis & Handling
pandas>=1.5.0
numpy>=1.23.0
openpyxl>=3.1.0

# Networking & Scraping
requests>=2.28.0
beautifulsoup4>=4.11.0
feedparser>=6.0.0
duckduckgo-search>=3.8.0

# Document Synthesis
reportlab>=3.6.0

# Utilities
python-dotenv>=1.0.0
python-dateutil>=2.8.0

# Optional Enhancements (Future Proofing)
yfinance>=0.2.0
scikit-learn>=1.2.0
matplotlib>=3.6.0
plotly>=5.10.0